Ordenado codigo y funciones
