+++ utilities/chats.py
@@ -50,3 +50,19 @@ def track_chats(update: Update, context: CallbackContext) -> None:
             logger.info("%s removed the bot from the channel %s",
                         cause_name, chat.title)
             context.bot_data.setdefault("channel_ids", set()).discard(chat.id)
+
+
+def show_chats(update: Update, context: CallbackContext) -> None:
+    """Shows which chats the bot is in"""
+    user_ids = ", ".join(str(uid)
+                         for uid in context.bot_data.setdefault("user_ids", set()))
+    group_ids = ", ".join(str(gid)
+                          for gid in context.bot_data.setdefault("group_ids", set()))
+    channel_ids = ", ".join(
+        str(cid) for cid in context.bot_data.setdefault("channel_ids", set()))
+    text = (
+        f"@{context.bot.username} is currently in a conversation with the user IDs {user_ids}."
+        f" Moreover it is a member of the groups with IDs {group_ids} "
+        f"and administrator in the channels with IDs {channel_ids}."
+    )
+    update.effective_message.reply_text(text)
